Fix Clang bytecode wire warning

+++ src/bytecode/bytecode_wire.c
@@ -1,7 +1,7 @@
 #include "bytecode_wire.h"
 
 uint16_t bc_wire_load_u16(const uint8_t *p) {
-  return (uint16_t)p[0] | ((uint16_t)p[1] << 8);
+  return (uint16_t)((uint16_t)p[0] | ((uint16_t)p[1] << 8));
 }
 
 uint32_t bc_wire_load_u32(const uint8_t *p) {
